requirements.txt
astroid==2.15.4
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
dill==0.3.6
Flask==2.2.3
Flask-Cors==3.0.10
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
platformdirs==3.2.0
pycparser==2.21
PyJWT==2.6.0
pylint==2.17.3
python-dotenv==1.0.0
six==1.16.0
tomlkit==0.11.7
Werkzeug==2.2.3
wrapt==1.15.0
